% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{geom_null_line}
\alias{geom_null_line}
\title{Dashed "no effect" reference line}
\usage{
geom_null_line(xintercept = 0, colour = "grey30")
}
\arguments{
\item{xintercept}{Numeric position of the line.}

\item{colour}{Line colour.}
}
\value{
A ggplot2 layer.
}
\description{
Small wrapper around \code{\link[ggplot2:geom_vline]{ggplot2::geom_vline()}} used by several stanviz
plots to mark a null reference.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + geom_point() + geom_null_line(3)
}
